iOS: Test the restore behaviour in the coordinator's own test file
Folds the tests into AIChatContextualSheetCoordinatorTests, reusing its
harness and doubles, so no new file and no project entry. Six cases through
presentSheet: a saved chat is restored, a deleted one is not, and the four
ways the store can fail to answer all still restore.
